Rename components and functions from AppImports* -> ImportedApps*
